Created input field to enter new todo to the todo list sans styling and functionality.
