+++ src/components/TodoComponents/TodoForm.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const TodoForm = props => {
+    return (
+        <div>
+            <input type="text" placeholder="What now?" ></input>
+        </div>
+    )
+}
+
+export default TodoForm;
